feature: ChildItem object now has String uid field, also updated child registration so their uid gets written into their node for the database

+++ app/src/main/java/com/example/smartair/ControllerLogs.java
@@ -2,7 +2,7 @@ package com.example.smartair;
 
 public class ControllerLogs {
 
-    public int dose;
+    public long dose;
     public long timestamp;
 
     public ControllerLogs() {}
