+++ app/src/main/java/com/example/smartair/RescueLogs.java
@@ -2,10 +2,10 @@ package com.example.smartair;
 
 public class RescueLogs {
 
-    public int dose;
-    public int preBreathRating;
-    public int postBreathRating;
-    public int postStatus;
+    public long dose;
+    public long preBreathRating;
+    public long postBreathRating;
+    public long postStatus;
     public long timestamp;
 
     public RescueLogs() {}
